update install script

+++ install.sh
@@ -1,13 +1,36 @@
 #!/bin/bash
-curl -fLo ~/.config/nvim/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+PLUG_VIM_URL=https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+curl -fLo ~/.config/nvim/autoload/plug.vim --create-dirs $PLUG_VIM_URL
 
-mkdir -p ~/.config/nvim/plugged
-mkdir -p ~/.vim
-ln -s ~/.config/nvim/autoload ~/.vim/autoload
-ln -s ~/.config/nvim/plugged ~/.vim/plugged
+if [ ! -d ~/.config/nvim ]
+then
+  mkdir -p ~/.config/nvim/plugged
+  mkdir -p ~/.vim
+  ln -s ~/.config/nvim/autoload ~/.vim/autoload
+  ln -s ~/.config/nvim/plugged ~/.vim/plugged
+  ln -s `pwd`/init.vim ~/.config/nvim/init.vim
+else
+  echo '--//warn: ~/.config/nvim directory is exist.'
+fi
 
-ln -s `pwd`/init.vim ~/.config/nvim/init.vim
-ln -s `pwd`/init.vim ~/.vimrc
-ln -s `pwd`/gitconfig ~/.gitconfig
-ln -s `pwd`/tmux.conf ~/.tmux.conf
-ln -s `pwd`/tmux.conf.local ~/.tmux.conf.local
+if [ ! -f ~/.vimrc ]
+then
+  ln -s `pwd`/init.vim ~/.vimrc
+else:
+  echo '--//warn: ~/.vimrc is exists.'
+fi
+
+if [ ! -f ~/.gitconfig ]
+then
+  ln -s `pwd`/gitconfig ~/.gitconfig
+else
+  echo '--//warn: ~/.gitconfig is exists.'
+fi
+
+if [ ! -f ~/.tmux.conf ]
+then
+  ln -s `pwd`/tmux.conf ~/.tmux.conf
+  ln -s `pwd`/tmux.conf.local ~/.tmux.conf.local
+else
+  echo '--//warn: ~/.tmux.conf is exists.'
+fi
